Persist task descriptions through sidebar edits
