+++ __tests__/e2e/tasks.spec.ts
@@ -439,7 +439,7 @@ test("task detail → shows correct fields after create", async ({ page }) => {
 
   await gotoNewTask(page, orgId);
   await page.getByLabel(/title/i).fill(taskTitle);
-  await page.getByLabel(/description/i).fill(description);
+  await page.locator("#description").fill(description);
   // Duration: 1h 30m
   await page.getByLabel("Hours").selectOption("1");
   await page.getByLabel("Minutes").selectOption("30");
@@ -449,7 +449,7 @@ test("task detail → shows correct fields after create", async ({ page }) => {
   await page.getByLabel(/min wait days/i).fill("2");
   await page.getByLabel(/max wait days/i).fill("5");
   await page.getByRole("button", { name: /create task/i }).click();
-  await expect(page).toHaveURL(`/orgs/${orgId}/tasks`, { timeout: 15000 });
+  await expect(page).toHaveURL(`/orgs/${orgId}/tasks`);
 
   // Navigate to detail
   await searchTasks(page, taskTitle);
@@ -501,21 +501,7 @@ test("task detail → shows updated values after edit", async ({ page }) => {
   await page.getByLabel("Minutes").selectOption("15");
   await page.getByLabel(/min wait days/i).fill("3");
   await page.getByLabel(/max wait days/i).fill("7");
-  // TipTap's ProseMirror pipeline is bypassed by Playwright's fill() and
-  // keyboard events in headless mode: the contenteditable DOM is updated but
-  // TipTap's onUpdate never fires, so the hidden <input name="description">
-  // keeps the old value. Set it directly after all other fields are filled
-  // (any earlier and React re-renders from other field changes can overwrite
-  // it by re-running TipTap's sync effect).
-  await page.evaluate(
-    (desc) => {
-      const input = document.querySelector<HTMLInputElement>(
-        'input[name="description"]',
-      );
-      if (input) input.value = desc;
-    },
-    updatedDescription,
-  );
+  await page.locator("#description").fill(updatedDescription);
   await page.getByRole("button", { name: /save/i }).click();
   await expect(page).toHaveURL(/\/orgs\/.+\/tasks\/.+(?<!\/edit)$/);
