+++ prisma/seeds/orgs/donut-shop-a/donut-shop-a-metadata.ts
@@ -1,0 +1,69 @@
+export const TASK_TAGS: Record<string, string[]> = {
+  "Open Shop Checklist": ["Daily Ops", "Opening"],
+  "Close Shop Checklist": ["Daily Ops", "Closing"],
+  "Mid-Day Stock Check": ["Daily Ops"],
+  "Restock Packaging & Supplies": ["Daily Ops"],
+  "Fryer Oil Quality Check": ["Fryer", "Quality"],
+  "Fry Morning Batches": ["Fryer"],
+  "Fry Afternoon Batches": ["Fryer"],
+  "Clean Fryer (End of Day)": ["Fryer", "Cleaning"],
+  "Quality Check \u2014 Display & Products": ["Quality", "Daily Ops"],
+  "Shift Handover": ["Daily Ops"],
+  "Make Custard Cream": ["Prep"],
+  "Make Choc Custard Cream": ["Prep"],
+  "Make Biscoff Filling": ["Prep"],
+  "Make Raspberry Cheesecake Filling": ["Prep"],
+  "Make Nutella Filling": ["Prep"],
+  "Make Peanut Butter Filling": ["Prep"],
+  "Prepare Classic Glaze": ["Prep"],
+  "Prepare Chocolate Fondant": ["Prep"],
+  "Prepare Biscoff Fondant": ["Prep"],
+  "Clean Fondant Bain-Marie": ["Prep", "Cleaning"],
+  "Recipe: White Choc Biscoff Frappe": ["Recipe"],
+  "Recipe: Honeycomb Frappe": ["Recipe"],
+  "Recipe: Coffee Frappe": ["Recipe"],
+  "Recipe: Salted Caramel Frappe": ["Recipe"],
+  "Recipe: Matcha Frappe": ["Recipe"],
+  "Recipe: Chocolate Milkshake": ["Recipe"],
+  "Recipe: Biscoff Custard Shake": ["Recipe"],
+  "Clean Ice Cream Machine": ["Cleaning"],
+  "Deep Clean Hatco (Hot Jam) Unit": ["Cleaning"],
+  "Deep Clean All Fridges": ["Cleaning"],
+  "Deep Clean Doughnut Display": ["Cleaning"],
+  "Clean & Tidy Storeroom": ["Cleaning"],
+};
+
+export const TASK_IMAGE_KEYWORDS: Record<string, string> = {
+  "Open Shop Checklist": "bakery,morning",
+  "Close Shop Checklist": "bakery,night",
+  "Mid-Day Stock Check": "bakery,shelf",
+  "Restock Packaging & Supplies": "packaging,boxes",
+  "Fryer Oil Quality Check": "frying,cooking oil",
+  "Fry Morning Batches": "doughnut,frying",
+  "Fry Afternoon Batches": "doughnut,cooking",
+  "Clean Fryer (End of Day)": "kitchen,cleaning",
+  "Quality Check \u2014 Display & Products": "doughnut,display",
+  "Shift Handover": "cafe,team",
+  "Make Custard Cream": "custard,cream",
+  "Make Choc Custard Cream": "chocolate,cream",
+  "Make Biscoff Filling": "caramel,cookie,spread",
+  "Make Raspberry Cheesecake Filling": "raspberry,cheesecake",
+  "Make Nutella Filling": "chocolate,hazelnut",
+  "Make Peanut Butter Filling": "peanut,spread",
+  "Prepare Classic Glaze": "doughnut,glaze,icing",
+  "Prepare Chocolate Fondant": "chocolate,fondant",
+  "Prepare Biscoff Fondant": "caramel,cookie",
+  "Clean Fondant Bain-Marie": "kitchen,pot,saucepan",
+  "Recipe: White Choc Biscoff Frappe": "white chocolate,frappe",
+  "Recipe: Honeycomb Frappe": "honeycomb,drink",
+  "Recipe: Coffee Frappe": "coffee,frappe",
+  "Recipe: Salted Caramel Frappe": "caramel,frappe",
+  "Recipe: Matcha Frappe": "matcha,green tea",
+  "Recipe: Chocolate Milkshake": "chocolate,milkshake",
+  "Recipe: Biscoff Custard Shake": "milkshake,caramel",
+  "Clean Ice Cream Machine": "ice cream,soft serve",
+  "Deep Clean Hatco (Hot Jam) Unit": "jam,kitchen",
+  "Deep Clean All Fridges": "refrigerator,fridge",
+  "Deep Clean Doughnut Display": "doughnut,display",
+  "Clean & Tidy Storeroom": "storage,shelves",
+};
