+++ prisma/seeds/orgs/donut-shop-a/donut-shop-a-tasks.ts
@@ -1,0 +1,335 @@
+export type TaskDef = [string, string, number, string, string, string, number, number];
+
+export const DONUT_TASKS: TaskDef[] = [
+  // ── Daily Operations ─────────────────────────────────────────────────────
+  [
+    "Open Shop Checklist",
+    "#F59E0B",
+    30,
+    "**Steps**\n1. Unlock front door and disable alarm.\n2. Turn on all lights and display cases.\n3. Power on fryer and preheat to 180°C.\n4. Set up POS terminal and float.\n5. Wipe down all counters and restock condiments.\n6. Check doughnut display stock levels and fill from overnight tray.\n7. Log opening time in shift register.",
+    "counter_staff",
+    "06:00",
+    0,
+    1,
+  ],
+  [
+    "Close Shop Checklist",
+    "#8B5CF6",
+    45,
+    "**Steps**\n1. Count and reconcile till. Record figures in shift register.\n2. Remove and label any remaining doughnuts for next-day staff meal.\n3. Turn off fryer — allow 30 min cool-down before cleaning.\n4. Wipe all surfaces, displays, and equipment exteriors.\n5. Mop floor (front of house and kitchen).\n6. Empty bins and replace liners.\n7. Set alarm and lock up.",
+    "shift_lead",
+    "17:00",
+    0,
+    1,
+  ],
+  [
+    "Mid-Day Stock Check",
+    "#22C55E",
+    20,
+    "**Steps**\n1. Count remaining doughnuts per flavour in display.\n2. Check frappe/shake ingredient levels (milk, ice, syrups, powders).\n3. Note any items running low and flag to manager.\n4. Restock from cool room as needed.\n5. Record stock status in the shift log.",
+    "counter_staff",
+    "12:00",
+    0,
+    1,
+  ],
+  [
+    "Restock Packaging & Supplies",
+    "#10B981",
+    25,
+    "**Check and restock:**\n• Doughnut boxes (individual, 6-pack, 12-pack)\n• Bags and tissue paper\n• Cups (8oz, 12oz, 16oz, 22oz)\n• Dome lids, flat lids, straw lids\n• Straws and soda spoons\n• Napkins\n• POS receipt paper\n\n_Reorder alert threshold: less than 1 full case of any item._",
+    "counter_staff",
+    "11:00",
+    1,
+    3,
+  ],
+  [
+    "Fryer Oil Quality Check",
+    "#EF4444",
+    15,
+    "**Steps**\n1. Check oil colour using test strip — replace if reading is above 25 TPM.\n2. Check oil level — top up if below fill line.\n3. Skim any debris from surface.\n4. Record result in equipment log.\n\n_Oil should be replaced every 3–4 days under normal volume. Do not fry in degraded oil._",
+    "fryer_op",
+    "07:30",
+    0,
+    2,
+  ],
+  [
+    "Fry Morning Batches",
+    "#EF4444",
+    60,
+    "**Steps**\n1. Confirm fryer is at 180°C.\n2. Remove proofed doughs from proofer.\n3. Lower rack gently — fry 90 sec each side.\n4. Drain on wire rack for 2 min.\n5. Cool completely before filling or glazing (min 20 min).\n6. Record batch count and any waste in production log.\n\n_Never overload the fryer — max 6 rings per side._",
+    "fryer_op",
+    "07:00",
+    0,
+    1,
+  ],
+  [
+    "Fry Afternoon Batches",
+    "#EF4444",
+    45,
+    "**Steps**\n1. Confirm fryer is still at 180°C (reheat if needed, 10 min).\n2. Fry top-up batches for afternoon/evening rush.\n3. Drain, cool, and pass to decorating station.\n4. Record batch count in production log.",
+    "fryer_op",
+    "13:00",
+    0,
+    1,
+  ],
+  [
+    "Clean Fryer (End of Day)",
+    "#EF4444",
+    40,
+    "**Steps**\n1. Allow oil to cool to below 50°C (check with probe).\n2. Drain oil into storage container — label with date.\n3. Wipe interior with paper towels.\n4. Fill with water + commercial fryer cleaner solution.\n5. Boil-out for 20 min.\n6. Drain, rinse twice with clean water.\n7. Dry thoroughly and reassemble.\n8. Record in equipment cleaning log.",
+    "fryer_op",
+    "17:30",
+    0,
+    1,
+  ],
+  [
+    "Quality Check — Display & Products",
+    "#A855F7",
+    20,
+    "**Steps**\n1. Inspect all displayed doughnuts — remove any that are stale, cracked, or poorly decorated.\n2. Check toppings are secure and glazes have set properly.\n3. Verify labels and allergen tags are correct.\n4. Taste test 1 item per flavour family (rotating schedule).\n5. Log any quality issues with photo if possible.",
+    "shift_lead",
+    "10:00",
+    0,
+    2,
+  ],
+  [
+    "Shift Handover",
+    "#64748B",
+    15,
+    "**Outgoing staff must:**\n1. Brief incoming staff on any ongoing issues.\n2. Note remaining stock levels verbally and in shift register.\n3. Flag any equipment issues or customer complaints.\n4. Hand over keys/float if applicable.\n5. Sign off shift register.",
+    "shift_lead",
+    "13:00",
+    0,
+    1,
+  ],
+
+  // ── Prep: Fillings ────────────────────────────────────────────────────────
+  [
+    "Make Custard Cream",
+    "#F59E0B",
+    30,
+    "**Ingredients**\n• 1250g Custard Powder\n• 2500ml Cold Water\n• 5000ml Cream\n\n**Method**\n1. Whisk cream and water together until combined.\n2. Fold in custard powder until smooth peaks form.\n\n_Makes approx. 8.75kg — enough for 215+ doughnuts. Should be light and fluffy, not dense._",
+    "fryer_op",
+    "06:30",
+    0,
+    1,
+  ],
+  [
+    "Make Choc Custard Cream",
+    "#F59E0B",
+    20,
+    "**Per 1kg Custard Cream:**\n• 10x small scoops Chocolate Powder\n\n**Method**\n1. Add Chocolate Powder to prepared Custard Cream.\n2. Mix thoroughly until fully incorporated.",
+    "fryer_op",
+    "06:45",
+    0,
+    1,
+  ],
+  [
+    "Make Biscoff Filling",
+    "#F59E0B",
+    15,
+    "**Ingredients**\n• 1000g Biscoff Spread\n• 40g Vegetable Oil\n\n**Method**\n1. Combine Biscoff and Vegetable Oil.\n2. Mix thoroughly.\n\n_Wet scoop with water before measuring Biscoff. Adding 4% Vegetable Oil ensures a workable consistency for filling._",
+    "fryer_op",
+    "07:00",
+    0,
+    2,
+  ],
+  [
+    "Make Raspberry Cheesecake Filling",
+    "#F59E0B",
+    20,
+    "**Per 1kg Custard Cream:**\n• 50g Quark\n• 2x small scoops crushed Freeze Dried Raspberries\n\n**Method**\n1. Add Quark and raspberries to prepared Custard Cream.\n2. Mix thoroughly.",
+    "fryer_op",
+    "07:00",
+    0,
+    2,
+  ],
+  [
+    "Make Nutella Filling",
+    "#F59E0B",
+    15,
+    "**Ingredients**\n• 3000g Nutella\n• 60g Vegetable Oil (2%)\n\n**Method**\n1. Add Vegetable Oil to Nutella.\n2. Mix until consistency is achieved — can take up to 5 minutes of hand mixing.\n\n_Wet scoop prior to use._",
+    "fryer_op",
+    "07:00",
+    0,
+    2,
+  ],
+  [
+    "Make Peanut Butter Filling",
+    "#F59E0B",
+    15,
+    "**Ingredients**\n• 1000g Peanut Butter\n• 200ml Vegetable Oil\n• 50g Icing Sugar _(NOT Snow Sugar)_\n\n**Method**\n1. Mix all ingredients thoroughly.\n\n_Makes enough for 100+ doughnuts._",
+    "fryer_op",
+    "07:00",
+    0,
+    2,
+  ],
+
+  // ── Prep: Glazes & Fondants ───────────────────────────────────────────────
+  [
+    "Prepare Classic Glaze",
+    "#EAB308",
+    15,
+    "Supplied from Bakery Group.\n\nMix all contents thoroughly before use. Heat gently to 60–65°C if too thick.",
+    "fryer_op",
+    "07:30",
+    0,
+    1,
+  ],
+  [
+    "Prepare Chocolate Fondant",
+    "#EAB308",
+    20,
+    "**Ingredients**\n• 1000g White Fondant\n• 100g Butter\n• 200g Chocolate Buttons\n• 60g Cocoa Powder\n• 60ml Hot Water\n\n**Method**\n1. Place all ingredients in bain-marie.\n2. Bring to 65°C while stirring continuously.",
+    "fryer_op",
+    "07:30",
+    0,
+    1,
+  ],
+  [
+    "Prepare Biscoff Fondant",
+    "#EAB308",
+    20,
+    "**Ingredients**\n• 1000g White Fondant\n• 200g Biscoff Spread\n\n**Method**\n1. Place all ingredients in bain-marie.\n2. Bring to 65°C while stirring.\n\n_Bain-marie requires 30+ min to heat adequately — plan ahead._",
+    "fryer_op",
+    "07:30",
+    0,
+    1,
+  ],
+  [
+    "Clean Fondant Bain-Marie",
+    "#EAB308",
+    30,
+    "**Steps**\n1. Turn off bain-marie, allow to cool 30 min.\n2. Remove pans — allow Fondants to set hard.\n3. Fill all Fondant pans (except Choc) with cold water, sit 20 min.\n4. Wipe sides and tops clean.\n5. Refill with fresh Fondant and return to clean bain-marie.",
+    "fryer_op",
+    "17:00",
+    0,
+    1,
+  ],
+
+  // ── Recipes: Frappes ─────────────────────────────────────────────────────
+  [
+    "Recipe: White Choc Biscoff Frappe",
+    "#8B5CF6",
+    5,
+    "**Ingredients**\n• 1 full cup Ice\n• 2/3 cup Milk\n• 1x large scoop Biscoff Spread\n• 4x small scoops White Chocolate Powder\n\n**Method**\n1. Blend 35 sec.\n2. Top with Whipped Cream Swirl and a dusting of Biscoff Crumb.\n\n_Wet the scoop with water before measuring Biscoff._",
+    "counter_staff",
+    "06:00",
+    0,
+    999,
+  ],
+  [
+    "Recipe: Honeycomb Frappe",
+    "#8B5CF6",
+    5,
+    "**Ingredients**\n• 1 full cup Ice\n• 2/3 cup Milk\n• 1.5x large scoops Honeycomb Frappe Powder\n• 1x large scoop Vanilla Frappe Powder\n• 12x Chocolate Buttons\n\n**Method**\n1. Blend 35 sec.\n2. Top with Whipped Cream Swirl and Dark Choc Flakettes.",
+    "counter_staff",
+    "06:00",
+    0,
+    999,
+  ],
+  [
+    "Recipe: Coffee Frappe",
+    "#8B5CF6",
+    5,
+    "**Ingredients**\n• 1 full cup Ice\n• 1/4 cup Milk\n• 1 double shot Espresso (60ml)\n• 4x small scoops Vanilla Frappe Powder\n\n**Method**\n1. Blend 35 sec.\n2. Top with Whipped Cream Swirl and Dark Chocolate Flakettes.",
+    "counter_staff",
+    "06:00",
+    0,
+    999,
+  ],
+  [
+    "Recipe: Salted Caramel Frappe",
+    "#8B5CF6",
+    5,
+    "**Ingredients**\n• 1 full cup Ice\n• 2/3 cup Milk\n• 3 pumps Salted Caramel Syrup (22.5ml)\n• 1x small scoop Salted Caramel Balls\n\n**Method**\n1. Blend 35 sec.\n2. Top with Whipped Cream Swirl and Silky Caramel lattice.",
+    "counter_staff",
+    "06:00",
+    0,
+    999,
+  ],
+  [
+    "Recipe: Matcha Frappe",
+    "#8B5CF6",
+    5,
+    "**Ingredients**\n• 1 full cup Ice\n• 2/3 cup Milk\n• 1x small scoop Matcha Powder\n\n**Method**\n1. Mix Matcha Powder with a splash of boiling water to form a paste first.\n2. Blend all 35 sec.\n3. Top with Whipped Cream Swirl and a dusting of Matcha Powder.\n\n_Always make paste fresh — no premix._",
+    "counter_staff",
+    "06:00",
+    0,
+    999,
+  ],
+
+  // ── Recipes: Milkshakes ───────────────────────────────────────────────────
+  [
+    "Recipe: Chocolate Milkshake",
+    "#EC4899",
+    5,
+    "**Small**\n• 1/2 cup Milk\n• 1/4 cup Soft Serve\n• 2 pumps Chocolate flavour\n\n**Large**\n• 1 cup Milk\n• 1/2 cup Soft Serve\n• 4 pumps Chocolate flavour\n\n**Thickshake** _(Large only)_\n• 3/4 cup Milk\n• 1 heaped cup Soft Serve\n• 4 pumps Chocolate flavour\n\n**Method**\n1. Blend 10 sec in metal cup.\n2. Serve in Striped cup with Slotted lid and straw.",
+    "counter_staff",
+    "06:00",
+    0,
+    999,
+  ],
+  [
+    "Recipe: Biscoff Custard Shake",
+    "#EC4899",
+    5,
+    "**Ingredients**\n• 1.5 cups Milk\n• 1.5 cups Ice\n• 1/2 cup Soft Serve\n• 1x large scoop Custard Powder\n• 1x large scoop Biscoff Spread\n\n**Method**\n1. Blend 20 sec.\n2. Top up with Milk if required.\n3. Serve in 22oz Striped cup with Slotted lid and straw.",
+    "counter_staff",
+    "06:00",
+    0,
+    999,
+  ],
+
+  // ── Weekly Cleaning ───────────────────────────────────────────────────────
+  [
+    "Clean Ice Cream Machine",
+    "#22C55E",
+    30,
+    "Full sanitize cycle. Scheduled **Monday** and **Friday**.\n\n_All cleaning tasks must be completed in full and signed off by those responsible._",
+    "counter_staff",
+    "14:00",
+    2,
+    4,
+  ],
+  [
+    "Deep Clean Hatco (Hot Jam) Unit",
+    "#22C55E",
+    45,
+    "Deep clean of the Hatco hot jam unit. Scheduled **Tuesday**.\n\n_All cleaning tasks must be completed in full and signed off by those responsible._",
+    "fryer_op",
+    "14:30",
+    5,
+    8,
+  ],
+  [
+    "Deep Clean All Fridges",
+    "#22C55E",
+    60,
+    "Deep clean interior and exterior of all fridges. Scheduled **Thursday**.\n\n_All cleaning tasks must be completed in full and signed off by those responsible._",
+    "shift_lead",
+    "14:00",
+    5,
+    8,
+  ],
+  [
+    "Deep Clean Doughnut Display",
+    "#22C55E",
+    30,
+    "Deep clean the doughnut display unit. Scheduled **Friday**.\n\n_All cleaning tasks must be completed in full and signed off by those responsible._",
+    "counter_staff",
+    "15:00",
+    5,
+    8,
+  ],
+  [
+    "Clean & Tidy Storeroom",
+    "#22C55E",
+    30,
+    "Clean and tidy the storeroom. Scheduled **Sunday**.\n\n_All cleaning tasks must be completed in full and signed off by those responsible._",
+    "shift_lead",
+    "15:00",
+    5,
+    8,
+  ],
+];
