+++ prisma/seeds/orgs/donut-shop-a/index.ts
@@ -1,0 +1,2 @@
+export { DONUT_TASKS, TASK_IMAGE_KEYWORDS, TASK_TAGS, type TaskDef } from "./data";
+export { seedDonutShopA } from "./donut-shop-a";
